site_builder building with lib.rs
